# docs/swagger.yaml
definitions:
  common.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.CommentCreateRequestSwaggo:
    properties:
      message:
        example: This is a comment message
        type: string
      photo_id:
        example: 1
        type: integer
    type: object
  model.CommentUpdateRequestSwaggo:
    properties:
      message:
        example: This is a comment message
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get all comments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Get all comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a new comment
      parameters:
      - description: Comment Create Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CommentCreateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Create a new comment
      tags:
      - Comments
  /comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by id
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Delete a comment by id
      tags:
      - Comments
    get:
      consumes:
      - application/json
      description: Get a comment by id
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Get a comment by id
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment by id
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      - description: Comment Update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CommentUpdateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Update a comment by id
      tags:
      - Comments
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"